pip-requirements.txt
Django
django-cors-headers
django-filter
djangorestframework
drf-writable-nested
asgiref
sqlparse